api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secure-rag-api"
version = "0.1.0"
description = "Secure multimodal RAG API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1",
  "uvicorn[standard]>=0.32.0,<1",
  "pydantic>=2.9.0,<3",
  "pydantic-settings>=2.6.0,<3",
  "httpx>=0.28.0,<1",
  "PyYAML>=6.0.2,<7",
  "SQLAlchemy>=2.0.35,<3",
  "psycopg[binary]>=3.2.0,<4",
  "qdrant-client>=1.11.0,<2",
  "llama-index-core==0.14.15",
  "llama-index-readers-google==0.7.2",
  "llama-index-embeddings-ollama==0.8.6",
  "llama-index-llms-ollama==0.9.1",
  "google-api-python-client>=2.140.0,<3",
  "google-auth>=2.30.0,<3",
  "google-auth-oauthlib>=1.2.0,<2",
  "python-jose[cryptography]>=3.3.0,<4",
  "prometheus-client>=0.21.0,<1",
  "rank-bm25>=0.2.2,<1",
  "pypdf>=5.0.0,<6",
  "python-docx>=1.1.2,<2",
  "Pillow>=10.4.0,<12",
  "PyMuPDF>=1.24.0,<2",
  "pytesseract>=0.3.13,<1",
  "orjson>=3.10.0,<4"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9",
  "pytest-asyncio>=0.24.0,<2",
  "pytest-cov>=5.0.0,<6",
  "respx>=0.22.0,<1"
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
